[#1]refactor: 에러 타입 변경, 핸들러 내용 수정
