+++ backend/src/main/java/com/prograngers/backend/entity/DataStructures.java
@@ -2,11 +2,8 @@ package com.prograngers.backend.entity;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
-import com.prograngers.backend.exception.notfound.AlgorithmNotFoundException;
-import com.prograngers.backend.exception.notfound.DataStructureNotFoundException;
+import com.prograngers.backend.exception.enumtype.DataStructureNotFoundException;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
 
 @AllArgsConstructor
 public enum DataStructures {
