+++ backend/src/main/java/com/prograngers/backend/exception/ErrorCode.java
@@ -4,7 +4,8 @@ public enum ErrorCode {
 
     INVALID_SOLUTION_BODY,
     SOLUTION_NOT_EXISTS,
-
-    ALGORITHM_NOT_EXISTS;
+    ALGORITHM_NOT_EXISTS,
+    DATASTRUCTURE_NOT_EXISTS,
+    LEVEL_NOT_EXISTS;
 
 }
