+++ backend/src/main/java/com/prograngers/backend/exception/enumtype/EnumTypeException.java
@@ -1,13 +1,13 @@
-package com.prograngers.backend.exception.notfound;
+package com.prograngers.backend.exception.enumtype;
 
 import com.prograngers.backend.exception.ErrorCode;
 import lombok.Getter;
 
 @Getter
-public class NotFoundException extends RuntimeException{
+public class EnumTypeException extends RuntimeException{
     ErrorCode errorCode;
 
-    public NotFoundException(String message, ErrorCode errorCode) {
+    public EnumTypeException(String message, ErrorCode errorCode) {
         super(message);
         this.errorCode = errorCode;
     }
